hubble/relay: split WithInsecure option for server and client
One needs to be able to allow Hubble Relay to establish its connections
to peers in the cluster over plaintext (`WithInsecureClient()`) and/or
to server over plaintext (`WithInsecureServer()`). Update the
`WithInsecure()` option to distinguish the two cases.

While here, rename internal variables in order to prevent confusion
between the two.

Signed-off-by: Robin Hahling <robin.hahling@gw-computing.net>